+++ packages/foam-vscode/src/core/services/markdown-parser.ts
@@ -7,15 +7,13 @@ import frontmatterPlugin from 'remark-frontmatter';
 import remarkDisableBlocks from 'remark-disable-tokenizers';
 import { parse as parseYAML } from 'yaml';
 import visit from 'unist-util-visit';
-import detectNewline from 'detect-newline';
-import os from 'os';
 import { NoteLinkDefinition, Resource, ResourceParser } from '../model/note';
 import { Position } from '../model/position';
 import { Range } from '../model/range';
-import { extractHashtags, extractTagsFromProp, isSome } from '../utils';
+import { extractHashtags, extractTagsFromProp, hash, isSome } from '../utils';
 import { Logger } from '../utils/log';
 import { URI } from '../model/uri';
-import { DisableMarkdownFeaturesSetting } from '../../settings';
+import { ICache } from '../utils/cache';
 
 export interface ParserPlugin {
   name?: string;
@@ -27,23 +25,38 @@ export interface ParserPlugin {
   onDidFindProperties?: (properties: any, note: Resource, node: Node) => void;
 }
 
-const ALIAS_DIVIDER_CHAR = '|';
+type Checksum = string;
+
+export interface ParserCacheEntry {
+  checksum: Checksum;
+  resource: Resource;
+}
+
+/**
+ * This caches the parsed markdown for a given URI.
+ *
+ * The URI identifies the resource that needs to be parsed,
+ * the checksum identifies the text that needs to be parsed.
+ *
+ * If the URI and the Checksum have not changed, the cached resource is returned.
+ */
+export type ParserCache = ICache<URI, ParserCacheEntry>;
 
 export function createMarkdownParser(
-  extraPlugins: ParserPlugin[],
-  disableMarkDownFeatures?: DisableMarkdownFeaturesSetting
+  extraPlugins: ParserPlugin[] = [],
+  cache?: ParserCache
 ): ResourceParser {
   const parser = unified()
     .use(markdownParse, { gfm: true })
     .use(frontmatterPlugin, ['yaml'])
-    .use(wikiLinkPlugin, { aliasDivider: ALIAS_DIVIDER_CHAR })
-    .use(remarkDisableBlocks, disableMarkDownFeatures);
+    .use(wikiLinkPlugin, { aliasDivider: '|' });
 
   const plugins = [
     titlePlugin,
     wikilinkPlugin,
     definitionsPlugin,
     tagsPlugin,
+    aliasesPlugin,
     sectionsPlugin,
     ...extraPlugins,
   ];
@@ -67,7 +80,6 @@ export function createMarkdownParser(
         }
       }
       const tree = parser.parse(markdown);
-      const eol = detectNewline(markdown) || os.EOL;
 
       const note: Resource = {
         uri: uri,
@@ -76,14 +88,9 @@ export function createMarkdownParser(
         title: '',
         sections: [],
         tags: [],
+        aliases: [],
         links: [],
         definitions: [],
-        source: {
-          text: markdown,
-          contentStart: astPointToFoamPosition(tree.position!.start),
-          end: astPointToFoamPosition(tree.position!.end),
-          eol: eol,
-        },
       };
 
       for (const plugin of plugins) {
@@ -101,12 +108,6 @@ export function createMarkdownParser(
               ...note.properties,
               ...yamlProperties,
             };
-            // Update the start position of the note by exluding the metadata
-            note.source.contentStart = Position.create(
-              node.position!.end.line! + 2,
-              0
-            );
-
             for (const plugin of plugins) {
               try {
                 plugin.onDidFindProperties?.(yamlProperties, note, node);
@@ -138,7 +139,23 @@ export function createMarkdownParser(
       return note;
     },
   };
-  return foamParser;
+
+  const cachedParser: ResourceParser = {
+    parse: (uri: URI, markdown: string): Resource => {
+      const actualChecksum = hash(markdown);
+      if (cache.has(uri)) {
+        const { checksum, resource } = cache.get(uri);
+        if (actualChecksum === checksum) {
+          return resource;
+        }
+      }
+      const resource = foamParser.parse(uri, markdown);
+      cache.set(uri, { checksum: actualChecksum, resource });
+      return resource;
+    },
+  };
+
+  return isSome(cache) ? cachedParser : foamParser;
 }
 
 /**
@@ -221,7 +238,10 @@ const sectionsPlugin: ParserPlugin = {
     }
   },
   onDidVisitTree: (tree, note) => {
-    const end = Position.create(note.source.end.line + 1, 0);
+    const end = Position.create(
+      astPointToFoamPosition(tree.position.end).line + 1,
+      0
+    );
     // Close all the remainig sections
     while (sectionStack.length > 0) {
       const section = sectionStack.pop();
@@ -255,6 +275,23 @@ const titlePlugin: ParserPlugin = {
   onDidVisitTree: (tree, note) => {
     if (note.title === '') {
       note.title = note.uri.getName();
+    }
+  },
+};
+
+const aliasesPlugin: ParserPlugin = {
+  name: 'aliases',
+  onDidFindProperties: (props, note, node) => {
+    if (isSome(props.alias)) {
+      const aliases = Array.isArray(props.alias)
+        ? props.alias
+        : props.alias.split(',').map(m => m.trim());
+      for (const alias of aliases) {
+        note.aliases.push({
+          title: alias,
+          range: astPositionToFoamRange(node.position!),
+        });
+      }
     }
   },
 };
@@ -306,7 +343,8 @@ const definitionsPlugin: ParserPlugin = {
     }
   },
   onDidVisitTree: (tree, note) => {
-    note.definitions = getFoamDefinitions(note.definitions, note.source.end);
+    const end = astPointToFoamPosition(tree.position.end);
+    note.definitions = getFoamDefinitions(note.definitions, end);
   },
 };