+++ packages/foam-vscode/src/extension.ts
@@ -1,4 +1,4 @@
-import { workspace, ExtensionContext, window } from 'vscode';
+import { workspace, ExtensionContext, window, commands } from 'vscode';
 import { MarkdownResourceProvider } from './core/services/markdown-provider';
 import { bootstrap } from './core/model/foam';
 import { URI } from './core/model/uri';
@@ -9,6 +9,10 @@ import { features } from './features';
 import { VsCodeOutputLogger, exposeLogger } from './services/logging';
 import { getIgnoredFilesSetting } from './settings';
 import { fromVsCodeUri, toVsCodeUri } from './utils/vsc-utils';
+import { AttachmentResourceProvider } from './core/services/attachment-provider';
+import { VsCodeWatcher } from './services/watcher';
+import { createMarkdownParser } from './core/services/markdown-parser';
+import VsCodeBasedParserCache from './services/cache';
 
 export async function activate(context: ExtensionContext) {
   const logger = new VsCodeOutputLogger();
@@ -27,28 +31,43 @@ export async function activate(context: ExtensionContext) {
       ['**/*'],
       getIgnoredFilesSetting().map(g => g.toString())
     );
+    const watcher = new VsCodeWatcher(
+      workspace.createFileSystemWatcher('**/*')
+    );
+    const parserCache = new VsCodeBasedParserCache(context);
+    const parser = createMarkdownParser([], parserCache);
+
     const markdownProvider = new MarkdownResourceProvider(
       matcher,
       dataStore,
-      triggers => {
-        const watcher = workspace.createFileSystemWatcher('**/*');
-        return [
-          watcher.onDidChange(uri => triggers.onDidChange(fromVsCodeUri(uri))),
-          watcher.onDidCreate(uri => triggers.onDidCreate(fromVsCodeUri(uri))),
-          watcher.onDidDelete(uri => triggers.onDidDelete(fromVsCodeUri(uri))),
-          watcher,
-        ];
-      }
+      parser,
+      watcher
+    );
+    const attachmentProvider = new AttachmentResourceProvider(
+      matcher,
+      dataStore,
+      watcher
     );
 
-    const foamPromise = bootstrap(matcher, dataStore, [markdownProvider]);
+    const foamPromise = bootstrap(matcher, dataStore, parser, [
+      markdownProvider,
+      attachmentProvider,
+    ]);
 
     // Load the features
     const resPromises = features.map(f => f.activate(context, foamPromise));
 
     const foam = await foamPromise;
-    Logger.info(`Loaded ${foam.workspace.list().length} notes`);
-    context.subscriptions.push(foam, markdownProvider);
+    Logger.info(`Loaded ${foam.workspace.list().length} resources`);
+    context.subscriptions.push(
+      foam,
+      watcher,
+      markdownProvider,
+      attachmentProvider,
+      commands.registerCommand('foam-vscode.clear-cache', () =>
+        parserCache.clear()
+      )
+    );
 
     const res = (await Promise.all(resPromises)).filter(r => r != null);