+++ packages/foam-vscode/src/test/test-utils.ts
@@ -9,6 +9,7 @@ import { FoamWorkspace } from '../core/model/workspace';
 import { Matcher } from '../core/services/datastore';
 import { MarkdownResourceProvider } from '../core/services/markdown-provider';
 import { NoteLinkDefinition, Resource } from '../core/model/note';
+import { createMarkdownParser } from '../core/services/markdown-parser';
 
 export { default as waitForExpect } from 'wait-for-expect';
 
@@ -22,10 +23,6 @@ export const TEST_DATA_DIR = URI.file(__dirname).joinPath(
 
 const position = Range.create(0, 0, 0, 100);
 
-const documentStart = position.start;
-const documentEnd = position.end;
-const eol = '\n';
-
 /**
  * Turns a string into a URI
  * The goal of this function is to make sure we are consistent in the
@@ -36,10 +33,15 @@ export const strToUri = URI.file;
 export const createTestWorkspace = () => {
   const workspace = new FoamWorkspace();
   const matcher = new Matcher([URI.file('/')], ['**/*']);
-  const provider = new MarkdownResourceProvider(matcher, {
-    read: _ => Promise.resolve(''),
-    list: _ => Promise.resolve([]),
-  });
+  const parser = createMarkdownParser();
+  const provider = new MarkdownResourceProvider(
+    matcher,
+    {
+      read: _ => Promise.resolve(''),
+      list: _ => Promise.resolve([]),
+    },
+    parser
+  );
   workspace.registerProvider(provider);
   return workspace;
 };
@@ -50,6 +52,7 @@ export const createTestNote = (params: {
   definitions?: NoteLinkDefinition[];
   links?: Array<{ slug: string } | { to: string }>;
   tags?: string[];
+  aliases?: string[];
   text?: string;
   sections?: string[];
   root?: URI;
@@ -68,6 +71,11 @@ export const createTestNote = (params: {
     tags:
       params.tags?.map(t => ({
         label: t,
+        range: Range.create(0, 0, 0, 0),
+      })) ?? [],
+    aliases:
+      params.aliases?.map(a => ({
+        title: a,
         range: Range.create(0, 0, 0, 0),
       })) ?? [],
     links: params.links
@@ -91,12 +99,6 @@ export const createTestNote = (params: {
               };
         })
       : [],
-    source: {
-      eol: eol,
-      end: documentEnd,
-      contentStart: documentStart,
-      text: params.text ?? '',
-    },
   };
 };